refactor(sidebar): extract company selector with react query (#346)
