+++ frontend/src/components/Sidebar.tsx
@@ -2,16 +2,9 @@ import { useEffect, useRef, useState } from 'react';
 import axios from 'axios';
 import { AnimatePresence, motion } from 'framer-motion';
 import { Link, NavLink } from 'react-router-dom';
-import api, { getApiErrorMessage } from '../api/client';
+import CompanySelector from './CompanySelector';
 import { useAuth } from '../context/AuthContext';
 import { useFY } from '../context/FYContext';
-import type {
-  CompanyCreationCapOut,
-  CompanyListItem,
-  CompanyProfile,
-  CompanyProfileUpdate,
-  CompanySelectOut,
-} from '../types/api';
 
 type NavItem = { to: string; label: string; end?: boolean };
 
@@ -63,20 +56,7 @@ export default function Sidebar({ isOpen = false, onClose }: SidebarProps) {
   const [newFYStartYear, setNewFYStartYear] = useState('');
   const [newFYError, setNewFYError] = useState('');
   const [newFYSubmitting, setNewFYSubmitting] = useState(false);
-  const [companyDropdownOpen, setCompanyDropdownOpen] = useState(false);
-  const [companyList, setCompanyList] = useState<CompanyListItem[]>([]);
-  const [companyLoading, setCompanyLoading] = useState(false);
-  const [companySwitchingId, setCompanySwitchingId] = useState<number | null>(null);
-  const [companyError, setCompanyError] = useState('');
-  const [canCreateCompany, setCanCreateCompany] = useState(true);
-  const [newCompanyModalOpen, setNewCompanyModalOpen] = useState(false);
-  const [newCompanyName, setNewCompanyName] = useState('');
-  const [newCompanySubmitting, setNewCompanySubmitting] = useState(false);
-  const [newCompanyError, setNewCompanyError] = useState('');
   const fyDropdownRef = useRef<HTMLDivElement>(null);
-  const companyDropdownRef = useRef<HTMLDivElement>(null);
-
-  const activeCompany = companyList.find((company) => company.is_active) ?? null;
 
   useEffect(() => {
     if (!fyDropdownOpen) return;
@@ -90,110 +70,11 @@ export default function Sidebar({ isOpen = false, onClose }: SidebarProps) {
   }, [fyDropdownOpen]);
 
   useEffect(() => {
-    if (!companyDropdownOpen) return;
-    const onClickOutside = (e: MouseEvent) => {
-      if (companyDropdownRef.current && !companyDropdownRef.current.contains(e.target as Node)) {
-        setCompanyDropdownOpen(false);
-      }
-    };
-    document.addEventListener('mousedown', onClickOutside);
-    return () => document.removeEventListener('mousedown', onClickOutside);
-  }, [companyDropdownOpen]);
-
-  useEffect(() => {
     if (!isOpen) return;
     const onKey = (e: KeyboardEvent) => { if (e.key === 'Escape') onClose?.(); };
     window.addEventListener('keydown', onKey);
     return () => window.removeEventListener('keydown', onKey);
   }, [isOpen, onClose]);
-
-  useEffect(() => {
-    if (!showCompanySwitcher) {
-      return;
-    }
-
-    if (!isAuthenticated) {
-      setCompanyList([]);
-      setCompanyError('');
-      setCanCreateCompany(false);
-      return;
-    }
-
-    let cancelled = false;
-    const loadCompanies = async () => {
-      setCompanyLoading(true);
-      setCompanyError('');
-      try {
-        const [companiesRes, capabilityRes] = await Promise.all([
-          api.get<CompanyListItem[]>('/company/companies'),
-          api.get<CompanyCreationCapOut>('/company/companies/capability'),
-        ]);
-        if (cancelled) return;
-        setCompanyList(companiesRes.data);
-        setCanCreateCompany(capabilityRes.data.can_create_company);
-      } catch (error) {
-        if (cancelled) return;
-        setCanCreateCompany(false);
-        setCompanyError(getApiErrorMessage(error, 'Failed to load companies'));
-      } finally {
-        if (!cancelled) {
-          setCompanyLoading(false);
-        }
-      }
-    };
-
-    void loadCompanies();
-    return () => {
-      cancelled = true;
-    };
-  }, [isAuthenticated, showCompanySwitcher]);
-
-  const handleCompanySwitch = async (companyId: number) => {
-    if (companySwitchingId !== null) return;
-    setCompanySwitchingId(companyId);
-    setCompanyError('');
-    try {
-      const res = await api.post<CompanySelectOut>(`/company/select/${companyId}`);
-      localStorage.setItem('active_company_id', String(res.data.active_company_id));
-      window.location.reload();
-    } catch (error) {
-      setCompanyError(getApiErrorMessage(error, 'Failed to switch company'));
-      setCompanySwitchingId(null);
-    }
-  };
-
-  const handleCreateCompany = async () => {
-    const trimmedName = newCompanyName.trim();
-    if (!trimmedName) {
-      setNewCompanyError('Company name is required.');
-      return;
-    }
-    setNewCompanySubmitting(true);
-    setNewCompanyError('');
-    const payload: CompanyProfileUpdate = {
-      name: trimmedName,
-      address: '',
-      gst: '',
-      phone_number: '',
-      currency_code: 'USD',
-      email: '',
-      website: '',
-      bank_name: '',
-      branch_name: '',
-      account_name: '',
-      account_number: '',
-      ifsc_code: '',
-    };
-
-    try {
-      const createRes = await api.post<CompanyProfile>('/company/companies', payload);
-      await handleCompanySwitch(createRes.data.id);
-    } catch (error) {
-      setNewCompanyError(getApiErrorMessage(error, 'Failed to create company'));
-    } finally {
-      setNewCompanySubmitting(false);
-    }
-  };
 
   return (
     <>
@@ -273,109 +154,7 @@ export default function Sidebar({ isOpen = false, onClose }: SidebarProps) {
 
         {/* FY section — above footer */}
         <div style={{ padding: '12px 10px', borderTop: '1px solid var(--sidebar-border)' }}>
-          {showCompanySwitcher && (
-            <>
-              <p className="sidebar__group-label" style={{ marginBottom: '6px' }}>Company</p>
-              <div ref={companyDropdownRef} style={{ position: 'relative', marginBottom: '12px' }}>
-                <button
-                  className="button button--ghost"
-                  style={{ width: '100%', textAlign: 'left', display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}
-                  onClick={() => setCompanyDropdownOpen((v) => !v)}
-                  aria-haspopup="listbox"
-                  aria-expanded={companyDropdownOpen}
-                >
-                  <span>{activeCompany?.name?.trim() || 'Select company'}</span>
-                  <span style={{ fontSize: '0.75rem', opacity: 0.6 }}>▾</span>
-                </button>
-                {companyDropdownOpen && (
-                  <div
-                    role="listbox"
-                    style={{
-                      position: 'absolute',
-                      left: 0,
-                      right: 0,
-                      top: 'calc(100% + 4px)',
-                      background: 'var(--bg-card-strong)',
-                      border: '1px solid var(--line-strong)',
-                      borderRadius: '0.5rem',
-                      boxShadow: '0 4px 24px rgba(0,0,0,0.45)',
-                      zIndex: 100,
-                      overflow: 'hidden',
-                      color: 'var(--text)',
-                    }}
-                  >
-                    {companyLoading && (
-                      <p style={{ padding: '0.5rem 0.75rem', fontSize: '0.85rem', opacity: 0.7 }}>Loading companies...</p>
-                    )}
-                    {!companyLoading && companyList.length === 0 && (
-                      <p style={{ padding: '0.5rem 0.75rem', fontSize: '0.85rem', opacity: 0.6 }}>No companies found</p>
-                    )}
-                    {companyList.map((company) => (
-                      <button
-                        key={company.id}
-                        role="option"
-                        aria-selected={company.is_active}
-                        style={{
-                          display: 'flex',
-                          alignItems: 'center',
-                          gap: '0.5rem',
-                          width: '100%',
-                          textAlign: 'left',
-                          padding: '0.5rem 0.75rem',
-                          background: 'none',
-                          border: 'none',
-                          cursor: companySwitchingId === null ? 'pointer' : 'wait',
-                          fontWeight: company.is_active ? 700 : 400,
-                          fontSize: '0.875rem',
-                          color: 'inherit',
-                        }}
-                        disabled={companySwitchingId !== null}
-                        onClick={() => {
-                          void handleCompanySwitch(company.id);
-                          setCompanyDropdownOpen(false);
-                        }}
-                      >
-                        <span style={{ width: '1rem' }}>{company.is_active ? '✓' : ''}</span>
-                        {company.name || `Company #${company.id}`}
-                      </button>
-                    ))}
-                    {canCreateCompany && (
-                      <>
-                        <hr style={{ margin: '0.25rem 0', border: 'none', borderTop: '1px solid var(--line)' }} />
-                        <button
-                          style={{
-                            display: 'block',
-                            width: '100%',
-                            textAlign: 'left',
-                            padding: '0.5rem 0.75rem',
-                            background: 'none',
-                            border: 'none',
-                            cursor: 'pointer',
-                            fontSize: '0.875rem',
-                            color: 'inherit',
-                            fontWeight: 500,
-                          }}
-                          onClick={() => {
-                            setCompanyDropdownOpen(false);
-                            setNewCompanyName('');
-                            setNewCompanyError('');
-                            setNewCompanyModalOpen(true);
-                          }}
-                        >
-                          + New Company
-                        </button>
-                      </>
-                    )}
-                  </div>
-                )}
-              </div>
-              {companyError && (
-                <p style={{ marginTop: '-6px', marginBottom: '8px', color: 'var(--error, #ef4444)', fontSize: '0.8rem' }}>
-                  {companyError}
-                </p>
-              )}
-            </>
-          )}
+          {showCompanySwitcher && <CompanySelector />}
 
           <p className="sidebar__group-label" style={{ marginBottom: '6px' }}>Financial Year</p>
           <div ref={fyDropdownRef} style={{ position: 'relative' }}>
@@ -483,59 +262,6 @@ export default function Sidebar({ isOpen = false, onClose }: SidebarProps) {
 
       {/* New FY Modal */}
       <AnimatePresence>
-        {showCompanySwitcher && newCompanyModalOpen && (
-          <motion.div
-            className="modal-overlay"
-            initial={{ opacity: 0 }}
-            animate={{ opacity: 1 }}
-            exit={{ opacity: 0 }}
-            transition={{ duration: 0.2 }}
-            onClick={(e) => { if (e.target === e.currentTarget) setNewCompanyModalOpen(false); }}
-          >
-            <motion.div
-              className="modal-panel"
-              role="dialog"
-              aria-modal="true"
-              aria-label="Create new company"
-              initial={{ scale: 0.95, opacity: 0 }}
-              animate={{ scale: 1, opacity: 1 }}
-              exit={{ scale: 0.95, opacity: 0 }}
-              transition={{ duration: 0.2 }}
-              style={{ maxWidth: '28rem' }}
-            >
-              <h2 style={{ marginBottom: '1.25rem', fontSize: '1.125rem', fontWeight: 700 }}>New Company</h2>
-              <div style={{ display: 'flex', flexDirection: 'column', gap: '1rem' }}>
-                <label style={{ display: 'flex', flexDirection: 'column', gap: '0.25rem', fontSize: '0.875rem', fontWeight: 500 }}>
-                  Company name
-                  <input
-                    className="input"
-                    value={newCompanyName}
-                    onChange={(e) => setNewCompanyName(e.target.value)}
-                    autoFocus
-                  />
-                </label>
-                <p style={{ margin: 0, fontSize: '0.8rem', opacity: 0.7 }}>
-                  This creates a blank company profile. You can complete GST, address, and bank details on the Company page.
-                </p>
-                {newCompanyError && <p style={{ margin: 0, color: 'var(--error, #ef4444)', fontSize: '0.85rem' }}>{newCompanyError}</p>}
-                <div style={{ display: 'flex', gap: '0.75rem', justifyContent: 'flex-end', marginTop: '0.5rem' }}>
-                  <button type="button" className="button button--ghost" onClick={() => setNewCompanyModalOpen(false)}>
-                    Cancel
-                  </button>
-                  <button
-                    type="button"
-                    className="button"
-                    disabled={newCompanySubmitting}
-                    onClick={() => { void handleCreateCompany(); }}
-                  >
-                    {newCompanySubmitting ? 'Creating…' : 'Create Company'}
-                  </button>
-                </div>
-              </div>
-            </motion.div>
-          </motion.div>
-        )}
-
         {newFYModalOpen && (
           <motion.div
             className="modal-overlay"
